==> notebooks/1 . Convert And Clean Whatsapp Data.ipynb <==
# Append System path for systematic imports
import sys
if (not('./' in sys.path)):
    sys.path.append('../')

#lets clean and transform the chat data
import pandas as pd
import os
filename="WhatsApp Chat with Pradyumn Y.txt"
data_path = '../data'
fpath_input = os.path.join(data_path,filename)
df=pd.read_csv(fpath_input,header=None,error_bad_lines=False,encoding='utf8')

df= df.drop(0)
df.rename(columns={0:'Date',1:'Chat'},inplace=True)
Message= df["Chat"].str.split("-", n = 1, expand = True) 
df['Date']=df['Date'].str.replace(",","") 
df['Time']=Message[0]
df['Text']=Message[1]
Message1= df["Text"].str.split(":", n = 1, expand = True) 
df['Text']=Message1[1]
df['Name']=Message1[0]
df=df.drop(columns=['Chat'])
df['Text']=df['Text'].str.lower()
df['Text'] = df['Text'].str.replace('<media omitted>','MediaShared')
df['Text'] = df['Text'].str.replace('this message was deleted','DeletedMsg')    
df['Name'] = df['Name'].astype('str')

from collections import defaultdict
def convert_to_chat_covs(df):
    conv_start = defaultdict(lambda :[])
    conv_end = defaultdict(lambda :[])
    count = 0
    for i,j in df[['Text','Name']].values:
        if(j==' Pradyumn Y'):
            count+=1
            conv_start[count].append(i)
        else:
            conv_end[count].append(i)
    return conv_start,conv_end

conv_start,conv_end = convert_to_chat_covs(df)

conv_dicts = pd.DataFrame({'col_1':conv_start,'col_2':conv_end})

def join_as(x):
    try:
        if(len(x)!=1):
            return ''.join(x)
        else:
            return str(x[0])
    except:
        return 'nan'
conv_dicts.dropna(subset=['col_1','col_2'],inplace=True)
conv_dicts['response_text'] = conv_dicts['col_2'].apply(join_as)
conv_dicts['input_text'] = conv_dicts['col_1'].apply(join_as)

conv_dicts.head()

conversation_data = conv_dicts[['input_text','response_text']]

conversation_data.to_csv(os.path.join(data_path,'conversation_data.csv'),index=False)

import pandas as pd
import re

def remove_hyperlinks(text):
    # Remove any URLs from the text using regex
    return re.sub(r'http\S+|www\S+|https\S+', '', text, flags=re.MULTILINE)

def clean_whatsapp_data(csv_file):
    # Read the CSV file
    df = pd.read_csv(csv_file)
    df.dropna(inplace=True)
    df = df.rename(columns={'input_text':'question','response_text':'response'})
    df['question'] = df['question'].apply(lambda x:x.replace('"',''))
    df['response'] = df['response'].apply(lambda x:x.replace('"',''))

    # Assuming you have a 'Message' column in your CSV file
    # Remove hyperlinks from the 'Message' column
    df['question'] = df['question'].apply(remove_hyperlinks)
    df['response'] = df['response'].apply(remove_hyperlinks)

    # Filter out rows containing media content
    # Update the pattern if you have a different format
    media_pattern = r'<Media omitted>$|<Media omitted>$|image omitted$|video omitted$|GIF omitted$|sticker omitted$|audio omitted$|document omitted$'
    df = df[~df['question'].str.contains(media_pattern)]
    df = df[~df['response'].str.contains(media_pattern)]

    # Write the cleaned data to a new CSV file
#     df.to_csv(output_file, index=False)
    return df

# Usage example
input_csv = os.path.join(data_path,'conversation_data.csv')
# output_csv = 'cleaned_whatsapp_data.csv'
cleaned_df = clean_whatsapp_data(input_csv)


cleaned_df

cleaned_df['text_field'] = 'prady:'+cleaned_df['question']+'. aarish: '+cleaned_df['response']

cleaned_df[['text_field']].to_csv(os.path.join(data_path,'text_fields.csv'),index=False)
