==> notebooks/2. Embeddings creation.ipynb <==
# We cleaned the data let us read it here now.

import os
import pandas as pd
data_path = '../data'
filename = "text_fields.csv"
fpath_input = os.path.join(data_path,filename)
df = pd.read_csv(fpath_input)

# !pip install openai
# import openai
import openai
# initialize openai API key
openai.api_key =  #platform.openai.com
# Initialize an embedding model
embed_model = "text-embedding-ada-002"

# lets create embeddings for these chats
from time import sleep
from tqdm import tqdm
embeddings = []
for i in tqdm(df.itertuples()):
    texts = i.text_field
    try:
        res = openai.Embedding.create(input=texts, engine=embed_model)
    except:
        done = False
        while not done:
            sleep(5)
            try:
                res = openai.Embedding.create(input=texts, engine=embed_model)
                done = True
            except:
                pass
    embeddings.append(res)

# use this to dump the chat encodings file
import pickle
with open('chat_embeddings.pkl', 'wb') as file:
    # A new file will be created
    pickle.dump(embeddings, file)
