==> notebooks/3. Pinecone Push Data.ipynb <==
import inspect

from getpass import getpass
from langchain import OpenAI
from langchain.chains import LLMChain, ConversationChain
from langchain.chains.conversation.memory import (ConversationBufferMemory, 
                                                  ConversationSummaryMemory, 
                                                  ConversationBufferWindowMemory,
                                                  ConversationKGMemory)
from langchain.callbacks import get_openai_callback
import tiktoken
import pickle
import os
import pandas as pd

data_path = '../data'
filepath = os.path.join(data_path,'chat_embeddings.pkl')
with open(filepath,'rb') as file:
    # A new file will be created
    chat_embeddings = pickle.load(file)

# Check the length of the embedding files
len(chat_embeddings[0]['data'][0]['embedding'])

import pinecone
pinecone_api_key= # place your key here
# initialize connection to pinecone
pinecone.init(
    api_key=pinecone_api_key,  # app.pinecone.io (console)
    environment="eu-west1-gcp"  # next to API key in console
)
# check if index already exists (it shouldn't if this is first time)

index_name = 'personalized-bot'
if index_name not in pinecone.list_indexes():
    # if does not exist, create index
    pinecone.create_index(
        index_name,
        dimension=len(chat_embeddings[0]['data'][0]['embedding']),
        metric='dotproduct'
    )

index = pinecone.Index(index_name)
# view index stats
index.describe_index_stats()

filename = 'text_fields.csv'
df = pd.read_csv(os.path.join(data_path,filename))

df

vectors= []
for vector in chat_embeddings:
    vectors.append(vector['data'][0]['embedding'])

text = df['text_field'].values
meta_batch = [{
        'text': x} for x in df['text_field'].values]
df['vector_id'] = df.index
df['vector_id'] = df['vector_id'].apply(str)
ids = (df['vector_id'].values)

from typing import Iterator
import numpy as np
class BatchGenerator:
    """ Models a simple batch generator that make chunks out of an input DataFrame. """
    
    def __init__(self, batch_size: int = 10) -> None:
        self.batch_size = batch_size
    
    def to_batches(self, df: pd.DataFrame) -> Iterator[pd.DataFrame]:
        """ Makes chunks out of an input DataFrame. """
        splits = self.splits_num(df.shape[0])
        if splits <= 1:
            yield df
        else:
            for chunk in np.array_split(df, splits):
                yield chunk
    
    def splits_num(self, elements: int) -> int:
        """ Determines how many chunks DataFrame contians. """
        return round(elements / self.batch_size)
    
    __call__ = to_batches

df_batcher = BatchGenerator(300)

df['vector_embeddings'] = vectors

df

# lets uspert the data using batches for better speed (note that the batch size set is 300)
k = []
for batch_df in df_batcher(df):
    meta_batch = [{
        "text": x} for x in batch_df['text_field'].values]
    k.append(list(zip(batch_df.vector_id, batch_df.vector_embeddings,meta_batch)))
    index.upsert(vectors=list(zip(batch_df.vector_id, batch_df.vector_embeddings,meta_batch)), namespace='index_1')
