==> notebooks/5. Langchain+ Pinecone + Openai.ipynb <==
import openai
from langchain import OpenAI, LLMChain, PromptTemplate
from langchain.chains import ConversationChain
from langchain.memory import ConversationBufferMemory


open_ai_key = #your open_ai_key goes here
openai.api_key = open_ai_key

import pinecone
# initialize connection to pinecone
pinecone_key = #your pinecone_key goes here
pinecone.init(
    api_key=pinecone_key,  # app.pinecone.io (console)
    environment="eu-west1-gcp"  # next to API key in console
)
index_name = 'personalized-bot'
index = pinecone.Index(index_name)

embed_model = "text-embedding-ada-002"
query = "Food preferences"
res = openai.Embedding.create(
    input=[query],
    engine=embed_model
)
xq = res['data'][0]['embedding']

result = index.query(xq, top_k=10, include_metadata=True,namespace='text_field_9')

contexts = [item['metadata']['text'] for item in result['matches']]

string1 = """
Using the provided conversation data between two persons, your task is to emulate the speaking style and language of one of the individuals while answering the user's question.
Draw upon your own knowledge and training to provide relevant and informative responses, but ensure that your answers reflect the mannerisms, tone, and language used by the chosen person. 
Always respond in the language of the conversation data shared, regardless of the question asked.
Chosen person is
"""

string2 = """
{chat_history}
human:{human_input}
Chatbot:""
"""

current_person = "Aarish"
primer = string1 + current_person + string2

prompt = PromptTemplate(
    input_variables=["chat_history","human_input"],
    template=primer
)

llm = OpenAI(openai_api_key=open_ai_key,temperature=0)

memory = ConversationBufferMemory(memory_key="chat_history")

llm_chain = LLMChain(
    llm=llm, 
    prompt=prompt, 
    verbose=True, 
    memory=memory,
)

input_query = 'kya khana hai?'
augmented_query = "\n\n---\n\n".join(contexts)+"\n\n-----\n\n"+ input_query

llm_chain.predict(human_input=augmented_query)

import sys
print(sys.executable)
